+++ src/tests.rs
@@ -11,8 +11,9 @@ use littlefs2::io::Result as LfsResult;
 use rand_core::{CryptoRng, RngCore};
 
 use crate::client::{CryptoClient as _, FilesystemClient as _};
+use crate::error::Error;
 use crate::types::{consent, reboot, ui, Bytes, Location, PathBuf};
-use crate::{api, block, platform, store, Error};
+use crate::{api, block, platform, store};
 
 pub struct MockRng(ChaCha20);
 
@@ -181,7 +182,7 @@ macro_rules! setup {
         let pc_interface: UserInterface = Default::default();
 
         let platform = $platform::new(rng, store, pc_interface);
-        let mut trussed: crate::Service<$platform> = crate::service::Service::new(platform);
+        let mut trussed: crate::service::Service<$platform> = crate::service::Service::new(platform);
 
         let (test_trussed_requester, test_trussed_responder) = crate::pipe::TRUSSED_INTERCHANGE
             .claim()
@@ -195,7 +196,7 @@ macro_rules! setup {
         trussed.set_seed_if_uninitialized(&$seed);
         let mut $client = {
             pub type TestClient<'a> =
-                crate::ClientImplementation<&'a mut crate::Service<$platform>>;
+                crate::client::ClientImplementation<&'a mut crate::service::Service<$platform>>;
             TestClient::new(test_trussed_requester, &mut trussed, None)
         };
     };